refactor: usefcmhandler 훅 분리
