+++ public/custom-sw.js
@@ -1,4 +1,3 @@
-// FCM 스크립트
 importScripts(
   "https://www.gstatic.com/firebasejs/10.12.1/firebase-app-compat.js",
 );
@@ -14,7 +13,6 @@ const messaging = firebase.messaging();
 messaging.onBackgroundMessage((payload) => {
   console.log("📩 Background message received:", payload);
 
-  //  notification or data에서 title/body 추출
   const title =
     payload.notification?.title || payload.data?.title || "📩 무비부키 알림";
   const body =
@@ -29,10 +27,8 @@ messaging.onBackgroundMessage((payload) => {
   });
 });
 
-// next-pwa의 워크박스 매니페스트
 self.__WB_MANIFEST;
 
-// 서비스워커 생명주기 관리
 self.addEventListener("install", (event) => {
   console.log("🔧 Service Worker installing...");
   self.skipWaiting();
