+++ src/app/_hooks/fcm/use-fcm.ts
@@ -1,4 +1,3 @@
-// hooks/useFCM.ts
 import { getToken, onMessage } from "firebase/messaging";
 import { getFirebaseMessaging } from "app/_lib/firebase-config";
 import { registerFCMToken } from "app/_apis/register-fcm-token";
@@ -29,22 +28,6 @@ export const useFCM = () => {
     }
   };
 
-  const getCurrentFCMToken = async () => {
-    try {
-      const messaging = await getFirebaseMessaging();
-      const registration = await navigator.serviceWorker.ready;
-
-      const token = await getToken(messaging!, {
-        vapidKey: process.env.NEXT_PUBLIC_FIREBASE_VAPID_KEY!,
-        serviceWorkerRegistration: registration,
-      });
-
-      return token;
-    } catch (err) {
-      console.error("❌ FCM 토큰 가져오기 실패:", err);
-      return null;
-    }
-  };
   const onForegroundMessage = (callback: (payload: any) => void) => {
     getFirebaseMessaging().then((messaging) => {
       console.log("onForegroundMessage 등록");
@@ -57,5 +40,5 @@ export const useFCM = () => {
     });
   };
 
-  return { requestPermissionAndToken, onForegroundMessage, getCurrentFCMToken };
+  return { requestPermissionAndToken, onForegroundMessage };
 };
